backend/requirements.txt
python-dotenv
websocket-client
pocketbase
pandas
pyspark
kafka-python